# metaclouds-backend-rust/docs/openapi-rust.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metaclouds Backend API",
    "description": "Rust 版 Metaclouds AI 算力调度后端，对齐 Go v1 API 契约。\n\n统一响应信封：`{\"success\":bool,\"data\":...,\"message\":...,\"code\":...,\"timestamp\":sec}`。错误响应对见 `ErrorResponse`。",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "development"
    }
  ],
  "paths": {
...
    "/api/v1/auth/change-password": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "`PUT /api/v1/auth/change-password` — 修改当前用户密码（需 JWT）。",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "password changed"
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "`GET /api/v1/auth/csrf` — 从 csrf_token Cookie 读取并返回双提交令牌。",
        "description": "未登录（无该 Cookie）时返回 401，引导前端重新登录。",
        "operationId": "get_csrf_token",
        "responses": {
          "200": {
            "description": "csrf token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "`POST /api/v1/auth/login`",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "login ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/profile": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "`GET /api/v1/auth/profile` — 返回当前登录用户（数据取自 DB）。",
        "operationId": "get_profile",
        "responses": {
          "200": {
            "description": "current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "`POST /api/v1/auth/refresh` — 用【尚未过期】的令牌换取一份新令牌。",
        "description": "路由挂在 `jwt_auth` 之后，因此只有仍有效的令牌能到达这里；\n这里用当前用户信息重新签发，刷新 Cookie 并返回 LoginResponse。",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "type": "object",
        "description": "`PUT /api/v1/auth/change-password` 请求体。",
        "required": [
          "old_password",
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string"
          },
          "old_password": {
            "type": "string"
          }
        }
      },
...
      "CsrfResponse": {
        "type": "object",
        "required": [
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "统一错误响应 schema（对齐 `ApiResponse` 错误形式）。\n\n实际信封：`{\"success\":false,\"message\":\"...\",\"code\":\"...\",\"timestamp\":<sec>}`。\n成功响应的 `data` 字段由各 handler 对应的业务 schema 描述。",
        "required": [
          "success",
          "message",
          "code",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "token",
          "user",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "integer",
            "format": "int64"
          },
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "description": "Public-facing user view (no password hash).",
        "required": [
          "id",
          "username",
          "email",
          "role",
          "tenant_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
}